[WIP] Create light docker image
Currently, our main Greenhouse-CI docker image is ~1GB, which is rather
large. In addition, the base docker image (RUBY) this is based on
suggests not to use the one we currentl use, but a light version: Alpine Ruby.

As such, this version is based of Alpine Ruby, which cuts the size to
~444MB. This should improve our pipeline performance.

This docker image still needs to be tested to ensure all appropriate
jobs will work on it.