Fix issue where `create-vsphere-from-vmx` fails to find stembuild
It appears that the `draft` option set on the `stembuild-release`
resource caused issue where it was unable to find any github releases.
The `draft` option is needed for posting new stembuild release, however.

As such, another stembuild release resource has been created, called
`post-stembuild-release`. This new one is designed for posting new draft
releases of stembuild.

The original release has been updated to no longer contain the draft
option, and is only used to check for releases of stembuild on github.

Co-authored-by: Jackson Feeny <jfeeny@pivotal.io>