switch stembuild-linux to use legacy-ci-image
- legacy-ci-image is the same as pivotalgreenhouse/ci, it's just pushed
to the pivnet dev registry as well as dockerhub
- only changed image in pipeline as individual tasks are more likely to
be run by hand (ie outside of pipeline) so adding the authentication to
each task is annoying [when running the tasks via fly execute].
- It does appear that streaming the docker image to each task adds a few
minutes slowdown. Hopefully the image is cached after the initial
download making each task initialization faster on subsequent runs.

[#175306351](https://www.pivotaltracker.com/story/show/175306351)
 Move integration test VMs to Nimbus