Debug long Azure publish step
We have a long (up to 7 days!) Azure Concourse task which doesn't
reliably complete. We suspect that the logic is flawed (" ... !=
'WaitForPublisherApproval'), but we can't be sure until we see the raw
JSON.

The raw JSON makes the output terribly chatty, and should be removed
once this problem is resolved.

Offer status format is documented here: https://docs.microsoft.com/en-us/azure/marketplace/cloud-partner-portal-api-retrieve-offer-status

[#177347625](https://www.pivotaltracker.com/story/show/177347625)

Co-authored-by: Brian Cunnie <bcunnie@vmware.com>