Automatically allow access to specific user groups
Previously there was a manual step to make the internal vSphere Windows
stemcell available to a list of particular groups. This is now baked
into the script.

This will have the side effect of granting access to the public release
to these user groups as well prior to publicizing the stemcell. However,
if they have access to the vSphere one it doesn't hurt to give them
access to the public IaaS versions ahead of time. The additional
available releases will also not affect their pipelines because it's a
separate product on PivNet.

[#177786296] Automate publishing of internal windows stemcell

Signed-off-by: Kenneth Lakin <klakin@vmware.com>