`promote-aws` job is resilient to AMI clean-up
The `promote-aws` job is not easily re-run, so we only want to fail if
we absolutely have to. And, to be honest, the
`cleanup-unpublished-aws-amis` is not terribly important--we've run for
years without cleaning them up. And since the script was new it was
buggy (it had at least two bugs in it).

This commit demotes the `cleanup-unpublished-aws-amis` task so it can no
longer cause the `promote-aws` job to fail.

Also, we leave a comment at the top of the pipeline's YAML file to
indicate how the pipeline should be flown.

[#180628580](https://www.pivotaltracker.com/story/show/180628580)

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>