Re-snapshot vsphere base images after installing updates
This should reduce the amount of time spent installing updates each time.

We also run dism.exe as part of this. This step takes ~45 minutes
when we run it later so having it as part of the snapshot should
speed things up. We still run it later, but it runs much faster due
to the early run.

We think the -dc flag in the new update-snapshot govc command is
redundant because we specfify the full path to the VM.
We kept the dc flag in anyway to keep consistent with our other
govc scripts.

Signed-off-by: Joseph Palermo <jpalermo@pivotal.io>
Co-authored-by: Joseph Palermo <jpalermo@pivotal.io>