Force CI image to rebuild on golang updates
The CI docker image uses the bosh/golang-release docker image to pull in
the latest version of Golang we're supporting. The CI docker image is
the one used to build the stembuild binary we ship to customers.

This change ensures that the image gets rebuilt each time a new version
of golang is available...