Add meta4 binary to ci-image
currently one task `tasks/commit-meta4-file/task.yml` requires meta4 and
it uses the `bosh/os-image-stemcell-publisher`.

this task appears to be the only one in `cloudfoundry` to use this
image.

this commit adds the `meta4` binary to the existing ci-image so that the
task can use this image instead of `bosh/os-image-stemcell-publisher`
which has a docker image in the bosh-linux-stemcell-builder repo