Builds from checked-out govmomi repo
Commit vmware/govmomi/commit/538dbef168288d18aaf55eeb044c6db38245a961
has utterly broken 'go install'. So, this commit changes things so that
we check out all of govmomi in a Concourse Resource and follow the
recommended steps of the govmomi maintainers to build the vcsim
executable that the test-contract-stembuild-linux Task requires.

We don't know that this will work, but hope that it will.

Signed-off-by: Kenneth Lakin <kenneth.lakin@broadcom.com>